Updated and corrected some things in nb_no.json
i have made the tips be in the right tense, corrected a fault in capitalizing, changed some translation to be better than what I first did, and added the use of « and ». It would be more grammatical correct for us than using ' ', even if we don't have them on our keyboards. Probably since we use the same keyboard layout in the Nordic, but Sweden and Denmark doesn't Quote the same way as us. Have also added an «\n» to the pre-alpha text to break up a line too long to be shown on the screen when the game is not in full screen.